implemented useRoute method
