+++ src/errors/routerRejectionError.ts
@@ -1,0 +1,11 @@
+import { RouterRejectionType } from '@/utilities'
+
+export class RouterRejectionError extends Error {
+  public type: RouterRejectionType
+
+  public constructor(type: RouterRejectionType) {
+    super()
+
+    this.type = type
+  }
+}
