+++ src/types/index.ts
@@ -1,4 +1,3 @@
-export * from './errors'
 export * from './hooks'
 export * from './injectionDepth'
 export * from './invalidRouteParamValueError'
