+++ src/utilities/hooks.ts
@@ -1,6 +1,5 @@
-import { NavigationAbortError } from '@/errors'
+import { NavigationAbortError, RouterPushError, RouterRejectionError } from '@/errors'
 import { RouteHookStore } from '@/models/RouteHookStore'
-import { RouterPushError, RouterRejectionError } from '@/types/errors'
 import { AfterRouteHook, AfterRouteHookResponse, BeforeRouteHook, BeforeRouteHookResponse, RouteHookAbort, RouteHookLifecycle } from '@/types/hooks'
 import { ResolvedRoute } from '@/types/resolved'
 import { RouterReject } from '@/types/router'
